opt: create Cluster type
